src/requirements_novo.txt
appdirs==1.4.3
certifi==2017.4.17
chardet==3.0.3
decorator==4.0.11
Django==1.11.13
django-appconf==1.0.2
django-compressor==2.1.1
django-crispy-forms==1.6.1
django-easy-select2==1.3.4
django-embed-video==1.1.2
django-extensions==2.0.7
django-filebrowser==3.9.1
django-floppyforms==1.7.0
django-grappelli==2.10.2
django-simple-captcha==0.5.5
django-simple-math-captcha==1.0.7
django-sortedm2m==1.3.3
django-suit==0.2.26
django-thumbnail-maker==0.0.5
django-widget-tweaks==1.4.2
freeze==1.0.10
gunicorn==19.7.1
idna==2.5
imagekit==1.0.6
ipython==6.0.0
ipython-genutils==0.2.0
jedi==0.10.2
olefile==0.44
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.1
prompt-toolkit==1.0.14
psycopg2==2.7.1
ptyprocess==0.5.1
Pygments==2.2.0
pyparsing==2.2.0
pytz==2017.2
rcssmin==1.0.6
requests==2.16.5
rjsmin==1.0.12
simplegeneric==0.8.1
six==1.10.0
-e git+https://github.com/mariocesar/sorl-thumbnail.git@66f0f355a976c76d712fdb83927175840c3a2478#egg=sorl_thumbnail
traitlets==4.3.2
urllib3==1.21.1
wcwidth==0.1.7
